requirements.txt
langchain 
langchain-google-community[calendar] 
google-api-python-client 
google-auth
python-dotenv
langchain-openai
langchain-anthropic
langchain-huggingface
langgraph
langgraph-checkpoint-postgres
psycopg[binary]
psycopg2-binary
dateparser
fastapi
uvicorn
streamlit
requests
pydantic
sentence-transformers
numpy
pendulum
colorama